% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotuj_dane.R
\name{przygotuj_dane}
\alias{przygotuj_dane}
\title{Przygotowanie danych z n wywolan gry.}
\usage{
przygotuj_dane(do_przygotowania, nazwa, n)
}
\arguments{
\item{do_przygotowania}{Wektor, ktory chcemy przygotowac.}

\item{nazwa}{Nazwa dla parametru, ktore opisuje wektor wejsciowy.}

\item{n}{Liczba wywolan gry.}
}
\description{
Funkcja przygotuj_dane() przygotowuje ramke daych z zadanego wektora z danymi.
}
\examples{
\dontrun{
dane_moc <- przygotuj_dane(moc,"Strategia_moc",100)
}

}

