% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabela_decyli.R
\name{tabela_decyli}
\alias{tabela_decyli}
\title{Zwraca tabele z decylami dla strategii.}
\usage{
tabela_decyli(dane1, dane2, dane3, strat)
}
\arguments{
\item{dane1}{Ramka danych z liczba ruchow dla gier z pierwsza strategia.}

\item{dane2}{Ramka danych z liczba ruchow dla gier z druga strategia.}

\item{dane3}{Ramka danych z liczba ruchow dla gier z trzecia strategia.}

\item{strat}{Nazwa strategii pierwszej.}
}
\description{
Funkcja tabela_decyli oblicza decyle liczby ruchow dla trzech strategii i zwraca je w tabeli.
}

