% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabela_statystyk.R
\name{tabela_statystyk}
\alias{tabela_statystyk}
\title{Generowanie tabeli z podstawowymi statystykami dla trzech strategii}
\usage{
tabela_statystyk(dane1, dane2, dane3, strat)
}
\arguments{
\item{dane1}{Ramka danych z liczba ruchow dla gier z pierwsza strategia.}

\item{dane2}{Ramka danych z liczba ruchow dla gier z druga strategia.}

\item{dane3}{Ramka danych z liczba ruchow dla gier z trzecia strategia.}

\item{strat}{Nazwa strategii pierwszej.}
}
\description{
Funkcja liczy podstawowe statystyki dla trzech strategii i przedstawia je w tabeli.
}

