% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wizytowka_SuperFarmer3.0_OlaAsia.R
\name{wizytowka_SuperFarmer3.0_OlaAsia}
\alias{wizytowka_SuperFarmer3.0_OlaAsia}
\title{Wizytowka do strategii z Pakietu SuperFarmer.SuperDziewczyn.}
\usage{
wizytowka_SuperFarmer3.0_OlaAsia(strategia, liczba_powtorzen = 1000)
}
\arguments{
\item{strategia}{Wybrana strategia, dla ktorej wizytowke chcemy uzyskac.}

\item{liczba_powtorzen}{Liczba wywolan gry celem uzyskania statystyk gier.}
}
\description{
Funkcja wizytowka() generuje wizytowke w formacie pdf opisujaca wybrana strategie z pakietu SuperFarmer.SuperDziewczyn. Oprocz wizualizacji w postaci wykresu gestosci dla 3 strategii oraz wykresu przebiegu jednej gry prezentujemy tabele statystyk i deycli dla trzech wybranych strategii.
}
\examples{
\dontrun{
wizytowka_strategia_owce <- wizytowka(SuperFarmer.SuperDziewczyn::strategia_owce)
}
}

