% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wykres_gestosci.R
\name{wykres_gestosci}
\alias{wykres_gestosci}
\title{Generowanie wykresu gestosci dla trzech strategii}
\usage{
wykres_gestosci(dane, strat, med, sr)
}
\arguments{
\item{dane}{Jest ramka danych zawierjaca trzy grupy dnaych do rysowania gestosci zawierajace nazwy strategii w kolumnie Strategia oraz liczby ruchow w pojedynczej grze w kolumnie Liczba_ruchow.}

\item{strat}{Nazwa strategii, ktora na wykresie porownujemy.}

\item{med}{Ramka danych z med liczby ruchow dla trzech strategii.}

\item{sr}{Ramka danych ze srednia liczby ruchow dla trzech strategii.}
}
\value{
Zwraca wykres gestosci.
}
\description{
Funkcja wykres_gestosci sluzy do tworzenia wykresu gestosci liczby ruchow dla trzech strategii, a takze nalozenia na niego
pionowych linii z med i srednia dla liczby ruchow dla kazdej strategii.
}

