% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wykres_zwierzat.R
\name{wykres_zwierzat}
\alias{wykres_zwierzat}
\title{Generowanie wykresu dla liczebnosci zwierzat podczas rozgrywki.}
\usage{
wykres_zwierzat(dane, tytul)
}
\arguments{
\item{dane}{Ramka danych zwierajaca liczebnosc zwiezat w rozgrywce. Musi zawierac kolumny Numer_kolejki, liczba, zwierze.}

\item{tytul}{Argument o klasie string definiujacy tytul wykresu.}
}
\description{
Funkcja wykres_zwierzat() sluzy do stworzenia wykresu punktowego, w ktorym punkty odpowiadaja liczbie zwierzat w pojedynczej rozgrywce w SuperFarmera.
}
\examples{
\dontrun{
wykres_male_zwierzeta <- wykres_wszytskie_zwierzatka(tabela_gra_owce,tytul_tabeli)
}

}
\seealso{
wykres_wszytskie_zwierzatka
}

