% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zapisz.R
\name{zapisz}
\alias{zapisz}
\title{Zapisanie obiektu do PDF.}
\usage{
zapisz(co, plik)
}
\arguments{
\item{co}{Definiowanie obiektu, ktory chce sie zapisac.}

\item{plik}{Nazwa pliku docelowego bez rozszerzenia.}
}
\description{
Funkcja zapisz zapisuje zadany obiekt do pliku PDF w folderze inst.
}

