% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wizytowka.R
\name{wizytowka}
\alias{wizytowka}
\title{Wizytowka do strategii z Pakietu SuperFarmer.SuperDziewczyn.}
\usage{
wizytowka(strategia, liczba_powtorzen = 1000)
}
\arguments{
\item{strategia}{Wybrana strategia, dla ktorej wizytowke chcemy uzyskac.}

\item{liczba_powtorzen}{Liczba wywolan gry celem uzyskania statystyk.}
}
\description{
Funkcja wizytowka generuje wizytowke w formacie pdf opisujaca i wizualizujaca wybrana strategie.
}
\examples{
\dontrun{
wizytowka_strategia_owce <- wizytowka(SuperFarmer.SuperDziewczyn::strategia_owce)
}

}

