% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wykres_wszystkie_zwierzatka.R
\name{wykres_wszystkie_zwierzatka}
\alias{wykres_wszystkie_zwierzatka}
\title{Wykres obrazujacy przebieg pojedynczej gry}
\usage{
wykres_wszystkie_zwierzatka(tabela, tytul)
}
\arguments{
\item{tabela}{tabela danych do wykresu}

\item{tytul}{tytul wykresu}
}
\value{
slupki wykres slupkowy
}
\description{
Wykres obrazujacy przebieg pojedynczej gry
}

